First launch of the website!
Today I finished coding my website for web design services. There are still things to be improved and worked on, and new features added, but for the sake of making my progress exactly 2 months from the 8th of July I decided to officially release it on 8th of September, which will be tomorrow. For testing purposes, I am firstly launching it here.